Move assert_shape_equal in utils, with assert_nD
